parallelize exporting
mexp is the slowest part of the pipeline.  Try to speed it up by
paralellizing it.

this adds a `pe' scripts that sits in front on N mexp childs and
dispatch threads to them, while still outputting the unchanged lines
for mkindex.

Unscientific testing has shown that this effectively reduces the time
for a full export, even if not linearly. (4 jobs cut the time in half,
8 jobs was just barely faster)