print a pagination before *and* after the listing
suggested by Stefan.

to do this, rework the way index pages are done.  Instead of
generating the first half and then appending the navigation and the
footer, generate only the body and change how fixfiles works to
prepend the heading.

this dance with temporary index files is needed because we don't know
how many pages there are until we've consumed all the input.