refactor the parsing into the module
change the mscan format string to siplify the parsing, and refactor
the parsing code into the module for reuse.

pe is an exception in that it doesn't care about the format string, it
just need to decide if the current line starts a new thread or not,
that's why it doesn't use the newly introduced `parse'.  The new
format simplifies pe too though.