build a cache of hashes for each thread
The idea is to use this cache to spot changed threads and re-export
all the mail of the threads that changed.  This will allow to add
thread navigation hints on each exported mail and have them up-to-date.

This makes the "skip if already exported" logic in mexp redundant
and incomplete.  pe will decide whether to export or not, and mexp
blindly do the work.